src/ml_strategies/random_forest_strategy.py
# Purpose: Implements Random Forest-based trading strategy
from sklearn.ensemble import RandomForestClassifier
import pandas as pd
import numpy as np
import logging

class RandomForestStrategy:
    def __init__(self, params):
        """Load Random Forest model and initialize parameters"""
        self.model = RandomForestClassifier(n_estimators=50, max_depth=5)
        self.symbol = params.get('symbol', 'EURUSD')
        self.features = ['rsi', 'macd']

    def train(self, data):
        """Train Random Forest model"""
        X = data[self.features]
        y = (data['close'].shift(-1) > data['close']).astype(int)
        self.model.fit(X[:-1], y[:-1])
        logging.info(f"Trained Random Forest model for {self.symbol}")

    def predict(self):
        """Implement ML-based signal generation logic"""
        from mt5_connector import MT5Connector
        mt5 = MT5Connector()
        if not mt5.initialize():
            return None

        data = mt5.fetch_market_data(self.symbol, mt5.TIMEFRAME_M15, 100)
        if data is None:
            return None

        data['rsi'] = ta.momentum.RSIIndicator(data['close'], window=14).rsi()
        macd = ta.trend.MACD(data['close'])
        data['macd'] = macd.macd()
        X = data[self.features].iloc[-1:].values

        prediction = self.model.predict(X)[0]
        if prediction == 1:
            logging.info(f"Random Forest Buy signal for {self.symbol}")
            return {'symbol': self.symbol, 'action': 'buy', 'volume': 0.1}
        return None